Wizard with multiple .steps and .actions sections.  Added wizard-header and wizard-footer for greater specificity of wizard parts.
Wrapped all content of a wizard inside the element with .wizard, whereas the previous version only required the "wizard" features (actions and steps) to be a part of the wrapper.  While this version worked, the steps, being tightly linked to the steps themselves, were not within the same hierarchy, thus making multiple .steps and .actions (as well as other features that go for the entire wizard) nearly impossible.