Add ability to pass error string back to datagrid
This adds the ability for the data source for a datagrid to signal
that an error occurred fetching the data and provide an error message
to display to the user.

As error is signaled by passing a string as the first argument to
the callback function instead of an object. The passed string is the
HTML to add to the table for the error message.