Spinner: up/down buttons shouldn't have tab index
when we are in a spinner text input and press tab, the focus is going to the spinner buttons, but this is not the same behavior of native jQuery spinner. To me, these buttons are useful only for mouse/touch/pointer, and not to keyboard, impairing the overall usability. So I did this patch to keep the same behaviour of native jQuery spinner.