Remove Creator section from README
I may have created the repository but you have all long surpassed the contributions I've made, and Fuel UX is a much bigger deal than I could have ever imagined! I appreciate the team leaving this credit in place for so long. This seems like as good a time as any to retire this section from the README.