Fixes #1311
Handling the addition of "bulk actions". Now when multiple items are selected returns all items with each items properties and the item itself.

Updating test for change in returned action obj

Make repeater actions have pointer cursor
Fixes #1411

Had to make a few changes to deal with the new additions within multi select. Added support for just selected instead of multi-selected.

Overwritting css class for actions dropdown menu to show glyphicons

Fix to index view. Column sizing on

Bad selector in multi-select-enabled if columnsyncing was not turned on

Adding a check to see if there are actions selected or not. If none selected bluk actions is disabled.
Center actions column and not using hard coded padding values

Updating the selecting of rows / clicking on rows.

Had to make a few changes to deal with the new additions within multi select. Added support for just selected instead of multi-selected.